+++ R/select_causal_rules.R
@@ -21,27 +21,30 @@ select_causal_rules <- function(rules_matrix_std, rules_list, ite_std,
 
   `%>%` <- magrittr::`%>%`
   rules <- NULL
-  if (stability_selection) {
-    # Stability selection
-    stab_mod <- stabs::stabsel(rules_matrix_std, ite_std,
-                               fitfun = "glmnet.lasso", cutoff = q,
-                               PFER = pfer_val, args.fitfun = "conservative")
-    rule_stab <- rules_list[stab_mod$selected]
-    select_rules <- rule_stab
+  if (length(rules_list)>1){
 
+    if (stability_selection) {
+      # Stability selection
+      stab_mod <- stabs::stabsel(rules_matrix_std, ite_std,
+                                 fitfun = "glmnet.lasso", cutoff = q,
+                                 PFER = pfer_val, args.fitfun = "conservative")
+      rule_stab <- rules_list[stab_mod$selected]
+      select_rules <- rule_stab
+
+    } else {
+      # LASSO
+      cv_lasso <- glmnet::cv.glmnet(rules_matrix_std, ite_std, alpha = 1,
+                                    intercept = FALSE)
+      aa <- stats::coef(cv_lasso, s = cv_lasso$lambda.1se)
+      index_aa <- which(aa[-1,1] != 0)
+      rule_LASSO <- data.frame(rules = rules_list[index_aa],
+                               val = aa[index_aa + 1, 1])
+      rule_LASSO <- rule_LASSO[order(-rule_LASSO[,2]), ] %>%
+        dplyr::filter(!is.na(rules))
+      select_rules <- rule_LASSO$rules
+    }
   } else {
-
-    # LASSO
-    cv_lasso <- glmnet::cv.glmnet(rules_matrix_std, ite_std, alpha = 1,
-                                  intercept = FALSE)
-    aa <- stats::coef(cv_lasso, s = cv_lasso$lambda.1se)
-    index_aa <- which(aa[-1,1] != 0)
-    rule_LASSO <- data.frame(rules = rules_list[index_aa],
-                             val = aa[index_aa + 1, 1])
-    rule_LASSO <- rule_LASSO[order(-rule_LASSO[,2]), ] %>%
-                  dplyr::filter(!is.na(rules))
-    select_rules <- rule_LASSO$rules
-
+    select_rules <- rules_list
   }
 
   return(select_rules)